+++ src/signatures.rs
@@ -1067,6 +1067,9 @@ signatures! {
     format = SonyDsdStreamFile
     value = b"DSD "
 
+    format = Squashfs
+    value = b"hsqs"
+
     format = TagImageFileFormat
     value = b"\x4D\x4D\x00\x2A"
     value = b"\x49\x49\x2A\x00"