+++ tests/application.rs
@@ -264,19 +264,6 @@ fn test_postscript() {
 }
 
 #[test]
-fn test_really_simple_syndication_1() {
-    let fmt = FileFormat::from_file("fixtures/application/sample1.rss").unwrap();
-    assert_eq!(fmt, FileFormat::ReallySimpleSyndication);
-}
-
-#[cfg(feature = "reader-xml")]
-#[test]
-fn test_really_simple_syndication_2() {
-    let fmt = FileFormat::from_file("fixtures/application/sample2.rss").unwrap();
-    assert_eq!(fmt, FileFormat::ReallySimpleSyndication);
-}
-
-#[test]
 fn test_realmedia() {
     let fmt = FileFormat::from_file("fixtures/application/sample.rm").unwrap();
     assert_eq!(fmt, FileFormat::Realmedia);