+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Squashfs
 - Windows Imaging Format (WIM)
 - openNURBS (3DM)
+- rzip (RZ)
 
 # Version 0.17.3 (2023-06-09)